Refactor to use hapi/glue to load handlers
