+++ package.json
@@ -22,6 +22,7 @@
   },
   "homepage": "https://github.com/zpratt/example-node-helm#readme",
   "dependencies": {
+    "@hapi/glue": "^7.0.0",
     "@hapi/hapi": "18.4.0",
     "@kubernetes/client-node": "^0.11.0"
   }
