+++ routes/health/index.js
@@ -1,0 +1,14 @@
+module.exports = {
+    name: 'health',
+    register: (server) => {
+        server.route({
+            method: 'GET',
+            path: '/health',
+            handler: async () => {
+                return {
+                    status: 'OK'
+                };
+            }
+        });
+    }
+};
